+++ src/copyConfigFile.js
@@ -2,37 +2,26 @@ const fs = require('fs')
 const path = require('path')
 const { promisify } = require('util')
 const copyFile = promisify(fs.copyFile)
-const writeFile = promisify(fs.writeFile)
-const eslintrc = require('./configFiles/.eslintrc.js')
+// const writeFile = promisify(fs.writeFile)
+
 // Copy config files
 let files = fs.readdirSync(path.resolve(__dirname, './configFiles'))
 
-module.exports = function (selectedArr) {
+module.exports = function ({type}) {
     console.log('Copying config files...')
-    console.log(files)
-
+    // console.log(files)
     let promArr = []
     for (const fileName of files) {
         // TODO: check if file exist
-        // not copy .eslintrc.js
-        if (fileName === '.eslintrc.js') continue
+        // not copy .eslintrc.xx.js
+        if (fileName.startsWith('.eslintrc')) continue
         let prom = copyFile(path.resolve(__dirname, 'configFiles', fileName), fileName)
         promArr.push(prom)
     }
 
-    // add special eslint config
-    if (selectedArr.includes('vue')) {
-        eslintrc.plugins.push('vue')
-        eslintrc.extends.splice(eslintrc.extends.length - 1, 0, 'plugin:vue/recommended') // prettier rule cover vue rule
-    }
-
-    if (selectedArr.includes('ts')) {
-        eslintrc.parserOptions.parser = '@typescript-eslint/parser'
-        eslintrc.plugins.push('@typescript-eslint')
-        eslintrc.extends.splice(eslintrc.extends.length - 1, 0, 'plugin:@typescript-eslint/recommended') // prettier rule cover this rule
-    }
     // create .eslintrc
-    promArr.push(writeFile('.eslintrc.js', 'module.exports = ' + JSON.stringify(eslintrc, null, 4)))
+    // promArr.push(writeFile('.eslintrc.js', 'module.exports = ' + JSON.stringify(eslintrc, null, 4)))
+    promArr.push(copyFile(path.resolve(__dirname, 'configFiles', `.eslintrc.${type}.js`), '.eslintrc.js'))
 
     return Promise.all(promArr).then(() => {
         console.log('√ Copying config files succeed')