+++ src/installNpmPkgs.js
@@ -1,11 +1,10 @@
 // install npm packages
 const childProcess = require('child_process')
 const eslintPkg = require('./version.json')
-module.exports = function (selectedArr) {
+module.exports = function ({type}) {
     let npmCmd = 'npm i -D'
-
-    for (const type of selectedArr) {
-        const pkgs = eslintPkg[type]
+    for (const key of ['default', type]) {
+        const pkgs = eslintPkg[key]
         for (const pkg in pkgs) {
             const version = pkgs[pkg]
             npmCmd += ' ' + pkg + '@' + version
@@ -13,7 +12,7 @@ module.exports = function (selectedArr) {
     }
     console.log('Installing npm packages...')
     console.log(npmCmd)
-    childProcess.exec(npmCmd, (err, stdout) => {
+    childProcess.exec(npmCmd, err => {
         if (err) return console.err('Npm install failed', err)
         console.log('√ Npm packages installation succeed')
         setTimeout(() => {